Package refactoring for test DataManager mocks
