+++ T03/app/src/androidTest/java/ca/ualberta/cmput301/t03/datamanager/CachedDataManagerTests.java
@@ -24,9 +24,10 @@ import com.google.gson.reflect.TypeToken;
 
 import java.io.IOException;
 import java.lang.reflect.Type;
-import java.util.HashMap;
 
 import ca.ualberta.cmput301.t03.common.exceptions.ServiceNotAvailableException;
+import ca.ualberta.cmput301.t03.datamanager.mocks.InMemoryDataManager;
+import ca.ualberta.cmput301.t03.datamanager.mocks.TestDto;
 
 import static ca.ualberta.cmput301.t03.common.ExceptionAsserter.assertThrowsException;
