+++ T03/app/src/androidTest/java/ca/ualberta/cmput301/t03/datamanager/JsonDataManagerTests.java
@@ -29,6 +29,7 @@ import java.lang.reflect.Type;
 import java.util.HashMap;
 
 import ca.ualberta.cmput301.t03.common.exceptions.NotImplementedException;
+import ca.ualberta.cmput301.t03.datamanager.mocks.TestDto;
 
 /**
  * Created by rishi on 15-10-29.
